Rust rustfmt, move tracing around.
